% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{check_and_install_dependencies}
\alias{check_and_install_dependencies}
\title{Check and Install Missing Dependencies}
\usage{
check_and_install_dependencies()
}
\description{
This function checks whether the required dependencies (\code{sangerseqR} and \code{Biostrings})
are installed on the user's system. If any of the packages are missing, it installs them
using \code{BiocManager}.
}
\details{
The function checks for the presence of the \code{BiocManager} package, and if it is not
installed, it installs it first. Then, it checks for the presence of \code{sangerseqR} and
\code{Biostrings}. If either of these packages is missing, the function installs them via
\code{BiocManager::install()}.
}
