% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Package Initialization}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{A character string representing the library name. This argument
is passed to \code{.onLoad} by the R environment when the package is loaded.}

\item{pkgname}{A character string representing the package name. This argument
is passed to \code{.onLoad} by the R environment when the package is loaded.}
}
\description{
This function is called when the package is loaded. It checks and installs
the required dependencies (\code{sangerseqR} and \code{Biostrings}) if they are not
already installed. It is automatically executed when the package is loaded.
}
\details{
The function first checks if \code{BiocManager} is installed. If it is not, it will
install it. Then, it checks for the presence of the required dependencies,
namely \code{sangerseqR} and \code{Biostrings}. If any of these packages are missing,
it installs them using \code{BiocManager}.
}
\seealso{
\code{\link{check_and_install_dependencies}} for the package installation logic.
}
